+++ NEWS.md
@@ -2,8 +2,10 @@
 
 * `GLPK` and `lpSolve` solvers are now supported
 * `useSolver()` function implemented, which calls solver APIs directly
-* `inspectSolution()` function implemented, which calls solver APIs directly
+* `inspectSolution()` and `appendSolution()` functions, which use `useSovler()` output
 * `prepareConstraints()` and `processGurobiOutput()` retired
+* `calculateIIF()` calculates item information function
+* `calculateExpectedRT()` calculates expected response times
 
 # eatATA 0.7.0